websocket/requirements.txt
opencv-python==4.8.1.78
flask==3.0.0
numpy==1.24.3
websockets==12.0
pillow==10.0.1 